updated jaxrs-cxf-non-spring-app sample

+++ samples/server/petstore/jaxrs-cxf-non-spring-app/src/test/java/io/swagger/api/PetApiTest.java
@@ -22,7 +22,6 @@
  * limitations under the License.
  */
 
-
 package io.swagger.api;
 
 import java.io.File;
@@ -49,7 +48,11 @@ import java.util.Map;
 
 
 /**
- * API tests for PetApi
+ * Swagger Petstore
+ *
+ * <p>This is a sample server Petstore server.  You can find out more about Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).  For this sample, you can use the api key `special-key` to test the authorization filters.
+ *
+ * API tests for PetApi 
  */
 public class PetApiTest {
 
@@ -68,11 +71,8 @@ public class PetApiTest {
         ClientConfiguration config = WebClient.getConfig(client); 
     }
 
-    
     /**
      * Add a new pet to the store
-     *
-     * 
      *
      * @throws ApiException
      *          if the Api call fails
@@ -80,32 +80,28 @@ public class PetApiTest {
     @Test
     public void addPetTest() {
         Pet body = null;
-	//api.addPet(body);
+        //api.addPet(body);
         
         // TODO: test validations
         
         
     }
-    
     /**
      * Deletes a pet
-     *
-     * 
      *
      * @throws ApiException
      *          if the Api call fails
      */
     @Test
     public void deletePetTest() {
-        Long petId = null;
+        Integer petId = null;
         String apiKey = null;
-	//api.deletePet(petId, apiKey);
+        //api.deletePet(petId, apiKey);
         
         // TODO: test validations
         
         
     }
-    
     /**
      * Finds Pets by status
      *
@@ -117,13 +113,12 @@ public class PetApiTest {
     @Test
     public void findPetsByStatusTest() {
         List<String> status = null;
-	//List<Pet> response = api.findPetsByStatus(status);
+        //List<Pet> response = api.findPetsByStatus(status);
         //assertNotNull(response);
         // TODO: test validations
         
         
     }
-    
     /**
      * Finds Pets by tags
      *
@@ -135,13 +130,12 @@ public class PetApiTest {
     @Test
     public void findPetsByTagsTest() {
         List<String> tags = null;
-	//List<Pet> response = api.findPetsByTags(tags);
+        //List<Pet> response = api.findPetsByTags(tags);
         //assertNotNull(response);
         // TODO: test validations
         
         
     }
-    
     /**
      * Find pet by ID
      *
@@ -152,18 +146,15 @@ public class PetApiTest {
      */
     @Test
     public void getPetByIdTest() {
-        Long petId = null;
-	//Pet response = api.getPetById(petId);
+        Integer petId = null;
+        //Pet response = api.getPetById(petId);
         //assertNotNull(response);
         // TODO: test validations
         
         
     }
-    
     /**
      * Update an existing pet
-     *
-     * 
      *
      * @throws ApiException
      *          if the Api call fails
@@ -171,51 +162,44 @@ public class PetApiTest {
     @Test
     public void updatePetTest() {
         Pet body = null;
-	//api.updatePet(body);
+        //api.updatePet(body);
         
         // TODO: test validations
         
         
     }
-    
     /**
      * Updates a pet in the store with form data
-     *
-     * 
      *
      * @throws ApiException
      *          if the Api call fails
      */
     @Test
     public void updatePetWithFormTest() {
-        Long petId = null;
+        Integer petId = null;
         String name = null;
         String status = null;
-	//api.updatePetWithForm(petId, name, status);
+        //api.updatePetWithForm(petId, name, status);
         
         // TODO: test validations
         
         
     }
-    
     /**
      * uploads an image
-     *
-     * 
      *
      * @throws ApiException
      *          if the Api call fails
      */
     @Test
     public void uploadFileTest() {
-        Long petId = null;
+        Integer petId = null;
         String additionalMetadata = null;
         org.apache.cxf.jaxrs.ext.multipart.Attachment file = null;
-	//ModelApiResponse response = api.uploadFile(petId, additionalMetadata, file);
+        //ModelApiResponse response = api.uploadFile(petId, additionalMetadata, file);
         //assertNotNull(response);
         // TODO: test validations
         
         
     }
-    
 }
